Refactor claim history presenter
