+++ app/presenters/claim_presenter.rb
@@ -1,5 +1,4 @@
 class ClaimPresenter < BasePresenter
-
   presents :claim
 
   def defendant_names
@@ -101,31 +100,5 @@ class ClaimPresenter < BasePresenter
 
   def assessment_value(assessment_attr)
     claim.assessment.new_record? ? h.number_to_currency(0) : h.number_to_currency(claim.assessment.__send__(assessment_attr))
-  end
-
-  def history_and_messages
-    dates = (messages.map(&:created_at) + versions.map(&:created_at)).compact.uniq.sort
-    date_strings = dates.map { |d| d.strftime(Settings.date_format) }.uniq
-
-    date_strings.inject({}) do |hash, date_string|
-      arr = select_by_date_string(versions, date_string) + select_by_date_string(messages, date_string)
-      arr = arr.flatten.sort_by { |e| e.created_at }
-      hash[date_string] = arr if arr.any?
-      hash
-    end
-  end
-
-  private
-
-  def messages
-    claim.messages.where('created_at IS NOT NULL').order(created_at: :asc)
-  end
-
-  def versions
-    claim.versions.order(created_at: :asc)
-  end
-
-  def select_by_date_string(collection, date_string)
-    collection.select { |e| e.created_at.strftime(Settings.date_format) == date_string }
   end
 end
