+++ spec/presenters/claim_presenter_spec.rb
@@ -164,50 +164,4 @@ RSpec.describe ClaimPresenter do
     claim.case_workers << FactoryGirl.build(:case_worker, user: FactoryGirl.build(:user, first_name: "Louis", last_name: 'Pasteur'))
     expect(subject.case_worker_names).to eq('Alexander Bell, Louis Pasteur')
   end
-
-  describe '#history_and_messages' do
-    let(:first_message) do
-      Timecop.travel(Time.zone.local(2015, 9, 21, 13, 0, 0)) do
-        create(:message, claim: claim, body: 'Hello world')
-      end
-    end
-
-    let(:second_message) do
-      Timecop.travel(Time.zone.local(2015, 9, 21, 14, 0, 0)) do
-        create(:message, claim: claim, body: 'Lorem ipsum')
-      end
-    end
-
-    let(:third_message) do
-      Timecop.travel(Time.zone.local(2015, 9, 23, 14, 0, 0)) do
-        create(:message, claim: claim, body: 'Lorem ipsum')
-      end
-    end
-
-    let!(:expected_hash) do
-      Timecop.travel(Time.zone.local(2015, 9, 21, 13, 50, 9)) do
-        claim.submit!
-      end
-
-      claim.reload
-
-      hash = {
-        '21/09/2015' => [
-          first_message,
-          claim.versions.last,
-          second_message
-        ],
-        '23/09/2015' => [
-          third_message
-        ]
-      }
-
-      hash
-    end
-
-    it 'returns an claims message and history hash in chronological order' do
-      claim.reload
-      expect(subject.history_and_messages).to eq(expected_hash)
-    end
-  end
 end
